refactor(variable): rename entity to entities
