+++ lib/src/features/variables/domain/entities/variable.dart
