+++ lib/src/features/variables/domain/repositories/variable_repository.dart
@@ -2,7 +2,7 @@
 
 import 'package:dartz/dartz.dart';
 import 'package:stadata_flutter_sdk/src/core/failures/failures.dart';
-import 'package:stadata_flutter_sdk/src/features/variables/domain/entity/variable.dart';
+import 'package:stadata_flutter_sdk/src/features/variables/domain/entities/variable.dart';
 import 'package:stadata_flutter_sdk/src/shared/domain/entities/api_response.dart';
 import 'package:stadata_flutter_sdk/stadata_flutter_sdk.dart';
