+++ lib/src/features/variables/domain/usecases/get_all_variables.dart
@@ -6,7 +6,7 @@ import 'package:injectable/injectable.dart';
 import 'package:stadata_flutter_sdk/src/base/usecase.dart';
 import 'package:stadata_flutter_sdk/src/core/di/service_locator.dart';
 import 'package:stadata_flutter_sdk/src/core/failures/failures.dart';
-import 'package:stadata_flutter_sdk/src/features/variables/domain/entity/variable.dart';
+import 'package:stadata_flutter_sdk/src/features/variables/domain/entities/variable.dart';
 import 'package:stadata_flutter_sdk/src/features/variables/domain/repositories/variable_repository.dart';
 import 'package:stadata_flutter_sdk/src/shared/domain/entities/api_response.dart';
