+++ test/src/features/variables/data/repositories/variable_repository_impl_test.dart
@@ -6,7 +6,7 @@ import 'package:stadata_flutter_sdk/src/core/typedef/typedef.dart';
 import 'package:stadata_flutter_sdk/src/features/variables/data/datasources/variable_remote_data_source.dart';
 import 'package:stadata_flutter_sdk/src/features/variables/data/models/variable_model.dart';
 import 'package:stadata_flutter_sdk/src/features/variables/data/repositories/variable_repository_impl.dart';
-import 'package:stadata_flutter_sdk/src/features/variables/domain/entity/variable.dart';
+import 'package:stadata_flutter_sdk/src/features/variables/domain/entities/variable.dart';
 import 'package:stadata_flutter_sdk/src/features/variables/domain/repositories/variable_repository.dart';
 import 'package:stadata_flutter_sdk/src/shared/data/models/api_response_model.dart';
 import 'package:stadata_flutter_sdk/src/shared/data/models/pagination_model.dart';
